Update database schema in MySQL Workbench
To ensure that the database schema is setup correctly even if the SQL
script is regenerated from the binary MySQL Workbench file, include
changes made to the SQL script into the MySQL Workbench file. Regenerate
the SQL script from MySQL Workbench to make sure that it is up to date.

Signed-off-by: Nicole Hoess <nicole.hoess@oth-regensburg.de>