Fix edge weights in database
When analysing developer communities and writing graphs to database,
extract edge weights directly from the cluster graph instead of manually
creating a weighted edge list.

Previously, the weighted edge list was generated by setting all edge
weights to one and then aggregating the edge weights from the same
source to the same destination. However, the weights in the edge list of
the input graph were already aggregated, meaning that there was a single
entry for each unique combination of source and destination. As a
consequence, the aggregated weights for these edges were overwritten by
ones. This caused all weights in the edgelist in the database to be set
to one as well, falsely implying that there was only one edge from
developer A to developer B.

Signed-off-by: Nicole Hoess <nicole.hoess@oth-regensburg.de>