Match identities with missing top level domain
Contributors may use e-mail addresses without a valid top-level domain,
e.g. example@localhost. In this case, the suffix "(none)" is appended to
the name and e-mail address of the contributor, resulting in them not
being matched correctly in the database. To prevent this, remove the
suffix to ensure correct matching based on the name of the contributor.

Signed-off-by: Nicole Hoess <nicole.hoess@oth-regensburg.de>