Update database configuration to prevent errors
When analysing larger projects, the database connection may timeout and
cause the entire analysis to crash. Changing the session timeout in the
code is not sufficient. Thus, we configure the database timeouts in
the docker entrypoint.

Spontaneously, the ID service may run into issues and become
unreachable. In this case, we can restart it manually, but may need more
time to intervene. Therefore, increase the retry count and wait time in
the ID manager.

When running multiple analysis scripts in parallel, it may happen that
the MySQL connection limit is exceeded. Therefore, we increase this
limit in the docker entrypoint.

Signed-off-by: Nicole Hoess <nicole.hoess@oth-regensburg.de>